Settings

+++ src/app/dashboard/dashboard.component.ts
@@ -5,6 +5,7 @@ import { MockDirectory } from '../shared/mocks/mock-directory';
 import { IndexedDbService } from '../indexedDb/indexed-db.service';
 import { ModalDirective } from 'ng2-bootstrap';
 import * as _ from 'lodash';
+import { Settings } from '../shared/models/settings';
 
 @Component({
   selector: 'app-dashboard',
@@ -16,9 +17,10 @@ export class DashboardComponent implements OnInit {
   workingDirectory: Directory;
   showCalculators: boolean = false;
   selectedCalculator: string;
-
+  // showSettings: boolean = false;
   isFirstChange: boolean = true;
   rootDirectoryRef: DirectoryDbRef;
+  view: string;
 
   @ViewChild('deleteModal') public deleteModal: ModalDirective;
   @ViewChild('deleteItemsModal') public deleteItemsModal: ModalDirective;
@@ -39,6 +41,13 @@ export class DashboardComponent implements OnInit {
               this.createDirectory();
             }
           })
+        this.indexedDbService.getDirectorySettings(1).then(
+          results => {
+            if (results.length == 0) {
+              this.createDirectorySettings();
+            }
+          }
+        );
       }
     )
   }
@@ -77,9 +86,17 @@ export class DashboardComponent implements OnInit {
   }
 
   viewCalculator(str: string) {
+    //this.showSettings = false;
     this.showCalculators = true;
     this.selectedCalculator = str;
   }
+
+  // viewSettings(bool: boolean) {
+  //   this.view = 'settings';
+  //   this.showSettings = true;
+  //   this.showCalculators = false;
+  //   this.selectedCalculator = '';
+  // }
 
   createExampleAssessments() {
     let tmpAssessment = MockDirectory.assessments[0];
@@ -93,6 +110,23 @@ export class DashboardComponent implements OnInit {
     this.indexedDbService.addAssessment(tmpAssessment).then(assessmentId => {
 
     })
+  }
+
+  createDirectorySettings() {
+    let tmpSettings: Settings = {
+      language: 'English',
+      currency: 'US Dollar',
+      unitsOfMeasure: 'Imperial',
+      directoryId: 1,
+      createdDate: new Date(),
+      modifiedDate: new Date()
+    }
+    this.indexedDbService.addSettings(tmpSettings).then(
+      results => {
+        console.log('root directory settings added');
+        console.log(results);
+      }
+    )
   }
 
 
@@ -143,8 +177,10 @@ export class DashboardComponent implements OnInit {
   }
 
   deleteSelected(dir: Directory) {
+    console.log('delete dir ' + dir.id);
     this.hideDeleteItemsModal();
     if (dir.subDirectory) {
+      console.log('if subDir ' + dir.id);
       dir.subDirectory.forEach(subDir => {
         if (subDir.delete || subDir.parentDirectoryId != 1) {
           this.indexedDbService.getChildrenDirectories(subDir.id).then(results => {
@@ -157,6 +193,7 @@ export class DashboardComponent implements OnInit {
       });
     }
     if (dir != this.workingDirectory) {
+      console.log('if != working ' + dir.id);
       this.indexedDbService.getDirectoryAssessments(dir.id).then(results => {
         let childDirAssessments = results;
         childDirAssessments.forEach(assessment => {
@@ -164,14 +201,31 @@ export class DashboardComponent implements OnInit {
             this.allDirectories = this.populateDirectories(this.rootDirectoryRef);
             this.workingDirectory = this.populateDirectories(this.workingDirectory);
           });
+          this.indexedDbService.getAssessmentSettings(assessment.id).then(
+            results => {
+              if (results.length != 0) {
+                this.indexedDbService.deleteSettings(results[0].id).then(
+                  results => { console.log('assessment setting deleter'); }
+                )
+              }
+            }
+          )
         })
+      })
+      if (dir.delete) {
         this.indexedDbService.deleteDirectory(dir.id).then(results => {
           this.allDirectories = this.populateDirectories(this.rootDirectoryRef);
           this.workingDirectory = this.populateDirectories(this.workingDirectory);
         })
 
-      })
-
+        this.indexedDbService.getDirectorySettings(dir.id).then(results => {
+          if (results.length != 0) {
+            this.indexedDbService.deleteSettings(results[0].id).then(
+              results => { console.log('dir setting deleted'); }
+            )
+          }
+        })
+      }
     }
     if (dir == this.workingDirectory) {
       let checkedAssessments = _.filter(this.workingDirectory.assessments, { 'delete': true });
@@ -180,6 +234,15 @@ export class DashboardComponent implements OnInit {
           this.allDirectories = this.populateDirectories(this.rootDirectoryRef);
           this.workingDirectory = this.populateDirectories(this.workingDirectory);
         });
+        this.indexedDbService.getAssessmentSettings(assessment.id).then(
+          results => {
+            if (results.length != 0) {
+              this.indexedDbService.deleteSettings(results[0].id).then(
+                results => { console.log('assessment setting deleter'); }
+              )
+            }
+          }
+        )
       })
     }
   }
